+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Improved console messages
 - Disable initial compile of crystal functions when starting Local
 - Fixed output bug when deploying from Local
+- Clearer message output for the command `confirm-account`
 
 ## 0.5.12 - Mar 16 2019
 - Add new branding to Local web templates.