+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Disable initial compile of crystal functions when starting Local
 - Fixed output bug when deploying from Local
 - Clearer message output for the command `confirm-account`
+- Ignore when Gemfile.lock is added
+- Don't trigger a deploy when Gemfile is added
+- Ignore Gemfile changes while initializing it
 
 ## 0.5.12 - Mar 16 2019
 - Add new branding to Local web templates.